Refactor: split classes
